+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - new: Add endpoint to review a place by a secret token
+- new: Add review tokens to reminder emails
 
 ## v0.12.1 (2023-01-04)